The BuildEventListener should be specific to Maven
